Run unit tests again
The `pear run-tests -r tests/` command executes the .phpt tests of which
there are none for Net/URL2:

    pear run-tests -r tests/
    Running 0 tests
    TOTAL TIME: 00:00
    0 PASSED TESTS
    0 SKIPPED TESTS

Running with `-u` for the Phpunit test-cases fails with the pear test
runner as long as it is not possible to install Phpunit via pear.

As an intermediate, use the phpunit test runner from the project
configuration (composer.json) and bootstrap across PHP versions.

Additionally, touch the test-cases for updates of PhpUnit since versions
3/4/5 up to 12.0.2 (tested with PH 5.3-8.4).